pyproject.toml
[tool.poetry]
name = "coursework-5"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pandas = "^2.2.3"
psycopg2 = "^2.9.10"
python-dotenv = "^1.1.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"


[tool.poetry.group.lint.dependencies]
mypy = "^1.15.0"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
